# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Robot Controller API",
    "description": "New backend service that provides resources for the Moon robot simulator.",
    "contact": {
      "name": "Kanishq Mehta",
      "email": "kanishq5001.be22@chitkara.edu.in"
    }
  },
  "paths": {
...
    "/api/robot-commands": {
      "get": {
        "tags": [
          "RobotCommands"
        ],
        "summary": "Retrieves all robot commands.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RobotCommand"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RobotCommand"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RobotCommand"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "RobotCommands"
        ],
        "summary": "Creates a robot command.",
        "description": "Sample request:\r\n            \r\n    POST /api/robot-commands\r\n    {\r\n       \"name\": \"DANCE\",\r\n       \"isMoveCommand\": true,\r\n       \"description\": \"Salsa on the Moon\"\r\n    }",
        "requestBody": {
          "description": "A new robot command from the HTTP request.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RobotCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RobotCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RobotCommand"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns the newly created robot command"
          },
          "400": {
            "description": "If the robot command is null",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "If a robot command with the same name already exists.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/robot-commands/{id}": {
      "get": {
        "tags": [
          "RobotCommands"
        ],
        "summary": "Retrieves robot commands by id.",
        "operationId": "GetRobotCommand",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "RobotCommands"
        ],
        "summary": "Updates a robot command.",
        "description": "Sample request:\r\n            \r\n    PUT /api/robot-commands\r\n    {\r\n       \"name\": \"DANCE\",\r\n       \"isMoveCommand\": true,\r\n       \"description\": \"Salsa on the Moon\"\r\n    }",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Updates the command with id.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "A updated robot command from the HTTP request.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RobotCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RobotCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RobotCommand"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns the newly created robot command"
          },
          "400": {
            "description": "If the robot command is null",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "If a robot command with the same name already exists.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "RobotCommands"
        ],
        "summary": "Deletes a robot command.",
        "description": "Sample request:\r\n    Post /api/robot-commands/1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Delete the robotcommand by id.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "202": {
            "description": "Deletes the command"
          },
          "404": {
            "description": "If the robot command is null"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "RobotCommand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isMoveCommand": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      }
    }
  }
}